test: split out serialization tests from business logic tests
