+++ bugsnag-android-core/src/test/java/com/bugsnag/android/NotifierTest.java
@@ -1,19 +1,11 @@
 package com.bugsnag.android;
 
-import static com.bugsnag.android.BugsnagTestUtils.streamableToJson;
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
-
-import androidx.test.filters.SmallTest;
-
-import org.json.JSONException;
-import org.json.JSONObject;
 import org.junit.Before;
 import org.junit.Test;
 
-import java.io.IOException;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
 
-@SmallTest
 public class NotifierTest {
 
     private Notifier notifier;
@@ -45,14 +37,5 @@ public class NotifierTest {
         String expected = "http://example.com";
         notifier.setURL(expected);
         assertEquals(expected, notifier.getURL());
-    }
-
-    @Test
-    public void testJsonSerialisation() throws JSONException, IOException {
-        JSONObject notifierJson = streamableToJson(notifier);
-        assertEquals(3, notifierJson.length());
-        assertEquals("Android Bugsnag Notifier", notifierJson.getString("name"));
-        assertNotNull(notifierJson.getString("version"));
-        assertEquals("https://bugsnag.com", notifierJson.getString("url"));
     }
 }
